Add files via upload
This code computes the BFS path based on the previous parent node list of a BFS tree. Initially I tried to work out the path starting from the start node. The problem is that I cannot adhere the goal node to the path list well at last. Then I changed my idea. Beginning with the goal node, I search its parent node and repeat such a process until finding the parent of the current node is the start node. then the path is returned. If the last element of my list is not the starting node after looking over all the elements in the parent node list, then the code will return the error message.